refactor(interfaces): move there VideoInterface
