+++ frontend/src/interfaces/VideoInterface.ts
@@ -1,0 +1,8 @@
+export interface VideoInterface {
+    id: number;
+    title: string;
+    url: string;
+    thumbnailUrl: string;
+    publish_date: string;
+    todays: boolean;
+  }
