+++ frontend/src/pages/IndexPage/IndexPage.tsx
@@ -1,24 +1,17 @@
 import { useState } from "react";
 import { useQuery } from "@tanstack/react-query";
 
-import axios from "../../api";
-
 import TodaysVideo from "../../components/Video/TodaysVideo";
 import VideoTitle from "../../components/Video/VideoTitle/VideoTitle";
 import CollectButton from "../../components/Video/CollectButton/CollectButton";
 import ErrorMessage from "../../components/forms/ErrorMessage";
 
+import { VideoInterface } from "../../interfaces/VideoInterface";
+
+import axios from "../../api";
+
 import pageStyles from "../Page.module.css";
 import styles from "./IndexPage.module.css";
-
-export interface VideoInterface {
-  id: number;
-  title: string;
-  url: string;
-  thumbnailUrl: string;
-  publish_date: string;
-  todays: boolean;
-}
 
 export const IndexPage = () => {
   const [showError, setShowError] = useState(false);
